C# Client Release 8.4.0

+++ AerospikeDemo/AsyncQuery.cs
@@ -1,5 +1,5 @@
 /* 
- * Copyright 2012-2020 Aerospike, Inc.
+ * Copyright 2012-2026 Aerospike, Inc.
  *
  * Portions may be licensed to Aerospike, Inc. under one or more contributor
  * license agreements.
@@ -54,7 +54,7 @@ namespace Aerospike.Demo
 
 			try
 			{
-				IndexTask task = client.CreateIndex(policy, args.ns, args.set, indexName, binName, IndexType.NUMERIC);
+				IndexTask task = client.CreateIndex(policy, args.ns, args.set, indexName, binName, IndexType.INTEGER);
 				task.Wait();
 			}
 			catch (AerospikeException ae)
